Added Component class, refactored containers
