+++ src/assets/factory/index.js
@@ -7,3 +7,11 @@ export const factorize = (data, container, ...callbacks) => {
     callback(data, builtContainer);
   }
 };
+
+export const createElement = (template) => {
+  const elementContainer = document.createElement(`div`);
+  elementContainer.insertAdjacentHTML(`beforeend`, template);
+
+  return elementContainer.firstElementChild;
+
+};
